--- a/native/bls_nif/src/lib.rs
+++ b/native/bls_nif/src/lib.rs
@@ -76,7 +76,7 @@ fn verify<'env>(public_key: Binary, message: Binary, signature: Binary) -> Resul
     }
     let sig = Signature::deserialize(signature.as_slice()).map_err(|err| format!("{:?}", err))?;
     let pubkey =
-        PublicKey::deserialize(public_key.as_slice()).map_err(|err| format!("{:?}", err))?;
+        fast_public_key_deserialize(public_key.as_slice()).map_err(|err| format!("{:?}", err))?;
 
     Ok(sig.verify(&pubkey, Hash256::from_slice(message.as_slice())))
 }
@@ -91,7 +91,7 @@ fn aggregate_verify<'env>(
         .map_err(|err| format!("{:?}", err))?;
     let pubkeys_result = public_keys
         .iter()
-        .map(|pkb| PublicKey::deserialize(pkb.as_slice()))
+        .map(|pkb| fast_public_key_deserialize(pkb.as_slice()))
         .collect::<Result<Vec<PublicKey>, _>>();
     let pubkeys = pubkeys_result.map_err(|err| format!("{:?}", err))?;
 
@@ -138,7 +138,7 @@ fn eth_fast_aggregate_verify<'env>(
         .map_err(|err| format!("{:?}", err))?;
     let pubkeys_result = public_keys
         .iter()
-        .map(|pkb| PublicKey::deserialize(pkb.as_slice()))
+        .map(|pkb| fast_public_key_deserialize(pkb.as_slice()))
         .collect::<Result<Vec<PublicKey>, _>>();
     let pubkeys = pubkeys_result.map_err(|err| format!("{:?}", err))?;
 
@@ -157,7 +157,7 @@ fn eth_aggregate_pubkeys<'env>(
         _ => {
             let pubkeys_result = public_keys
                 .iter()
-                .map(|pkb| PublicKey::deserialize(pkb.as_slice()))
+                .map(|pkb| fast_public_key_deserialize(pkb.as_slice()))
                 .collect::<Result<Vec<PublicKey>, _>>();
 
             let pubkeys = pubkeys_result.map_err(|err| format!("{:?}", err))?;
@@ -175,7 +175,7 @@ fn eth_aggregate_pubkeys<'env>(
 #[rustler::nif]
 fn key_validate<'env>(public_key: Binary) -> Result<bool, String> {
     let _pubkey =
-        PublicKey::deserialize(public_key.as_slice()).map_err(|err| format!("{:?}", err))?;
+        fast_public_key_deserialize(public_key.as_slice()).map_err(|err| format!("{:?}", err))?;
 
     Ok(true)
 }