@ElFantasma 2025-03-06
Note:

The validation included in `PublicKey::deserialize` is just a check that the key is on the eliptic curve:

```rust
            pub fn validate(&self) -> Result<(), BLST_ERROR> {
                unsafe {
                    if $pk_is_inf(&self.point) {
                        return Err(BLST_ERROR::BLST_PK_IS_INFINITY);
                    }
                    if !$pk_in_group(&self.point) {
                        return Err(BLST_ERROR::BLST_POINT_NOT_IN_GROUP);
                    }
                }
                Ok(())
            }
```
(As well as the Infinity Publick Check that we are still doing)
Avoiding this check does not prevent the aggregated signature verification to fail if some key is invalid.